@@ -1,6 +1,16 @@
 package com.thejawnpaul.gptinvestor.features.company.domain.model
 
-sealed class SectorInput {
-    data object AllSector : SectorInput()
-    data class CustomSector(val sectorName: String, val sectorKey: String) : SectorInput()
+sealed interface SectorInput {
+    val hasImage: Boolean
+
+    data object AllSector : SectorInput {
+        override val hasImage: Boolean
+            get() = false
+    }
+
+    data class CustomSector(
+        val sectorName: String,
+        val sectorKey: String,
+        override val hasImage: Boolean = false
+    ) : SectorInput
 }
