@@ -4,10 +4,12 @@ import android.net.Uri
 import androidx.compose.animation.AnimatedContent
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.IntrinsicSize
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -17,20 +19,12 @@ import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.KeyboardArrowDown
-import androidx.compose.material.icons.filled.KeyboardArrowUp
 import androidx.compose.material3.DividerDefaults
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.HorizontalDivider
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.OutlinedCard
 import androidx.compose.material3.PrimaryTabRow
-import androidx.compose.material3.SegmentedButton
-import androidx.compose.material3.SegmentedButtonDefaults
-import androidx.compose.material3.SingleChoiceSegmentedButtonRow
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Tab
 import androidx.compose.material3.Text
@@ -39,22 +33,22 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.SpanStyle
 import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextDecoration
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import coil.compose.AsyncImage
-import com.halilibo.richtext.commonmark.Markdown
-import com.halilibo.richtext.ui.material3.RichText
 import com.thejawnpaul.gptinvestor.R
 import com.thejawnpaul.gptinvestor.core.utility.toReadable
 import com.thejawnpaul.gptinvestor.core.utility.toTwoDecimalPlaces
@@ -62,97 +56,48 @@ import com.thejawnpaul.gptinvestor.features.company.data.remote.model.CompanyDet
 import com.thejawnpaul.gptinvestor.features.company.data.remote.model.HistoricalData
 import com.thejawnpaul.gptinvestor.features.company.presentation.model.NewsPresentation
 import com.thejawnpaul.gptinvestor.features.company.presentation.state.TimePeriod
-import com.thejawnpaul.gptinvestor.ui.theme.GPTInvestorTheme
+import com.thejawnpaul.gptinvestor.theme.GPTInvestorTheme
+import com.thejawnpaul.gptinvestor.theme.LocalGPTInvestorColors
+import com.thejawnpaul.gptinvestor.theme.linkMedium
 
 @Composable
-fun CompanyDetailDataSource(modifier: Modifier = Modifier, list: List<NewsPresentation> = emptyList(), source: String) {
-    var expanded by rememberSaveable { mutableStateOf(false) }
+fun CompanyDetailDataSource(modifier: Modifier, list: List<NewsPresentation> = emptyList(), onClick: () -> Unit) {
+    Surface(
+        modifier = modifier,
+        shape = RoundedCornerShape(12.dp),
+        border = BorderStroke(width = 2.dp, color = DividerDefaults.color),
+        onClick = onClick
+    ) {
+        Row(
+            modifier = Modifier.padding(12.dp),
+            verticalAlignment = Alignment.CenterVertically,
+            horizontalArrangement = Arrangement.spacedBy(8.dp)
+        ) {
+            Text(
+                text = stringResource(R.string.sources),
+                style = MaterialTheme.typography.bodySmall
+            )
+            OverlappingIcons(
+                modifier = Modifier,
+                images = list.map { it.imageUrl }.take(4)
+            )
+        }
+    }
+}
 
-    OutlinedCard(modifier = Modifier.padding(top = 16.dp, start = 16.dp, end = 16.dp)) {
-        if (expanded) {
-            Box(
+@Composable
+fun OverlappingIcons(modifier: Modifier, images: List<String>) {
+    Box(modifier) {
+        images.reversed().forEachIndexed { index, it ->
+            AsyncImage(
+                model = it,
                 modifier = Modifier
-                    .padding(8.dp)
-                    .fillMaxWidth()
-            ) {
-                RichText(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(vertical = 8.dp)
-                ) {
-                    Markdown(content = source)
-                }
-
-                IconButton(modifier = Modifier.align(Alignment.TopEnd), onClick = {
-                    expanded = !expanded
-                }) {
-                    Icon(
-                        imageVector = if (expanded) Icons.Filled.KeyboardArrowUp else Icons.Filled.KeyboardArrowDown,
-                        contentDescription = if (expanded) {
-                            stringResource(R.string.show_less)
-                        } else {
-                            stringResource(R.string.show_more)
-                        }
-                    )
-                }
-            }
-        } else {
-            Row(
-                modifier = Modifier
-                    .padding(8.dp)
-                    .fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceBetween,
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Row(verticalAlignment = Alignment.CenterVertically) {
-                    Image(
-                        modifier = Modifier.padding(end = 4.dp),
-                        painter = painterResource(R.drawable.data_from_icon),
-                        contentDescription = null
-                    )
-
-                    Text(
-                        text = stringResource(R.string.data_from),
-                        style = MaterialTheme.typography.titleMedium
-                    )
-                }
-
-                Row(verticalAlignment = Alignment.CenterVertically) {
-                    Row(verticalAlignment = Alignment.CenterVertically) {
-                        list.take(4).forEachIndexed { i, it ->
-                            Surface(
-                                modifier = Modifier
-                                    .size(24.dp)
-                                    .padding(),
-                                shape = CircleShape,
-                                border = BorderStroke(width = 2.dp, color = Color(227, 239, 252))
-                            ) {
-                                AsyncImage(
-                                    model = it.imageUrl,
-                                    contentDescription = null,
-                                    modifier = Modifier
-                                        .size(16.dp)
-                                        .align(Alignment.CenterVertically),
-                                    contentScale = ContentScale.Crop
-                                )
-                            }
-                        }
-                    }
-
-                    IconButton(onClick = {
-                        expanded = !expanded
-                    }) {
-                        Icon(
-                            imageVector = if (expanded) Icons.Filled.KeyboardArrowUp else Icons.Filled.KeyboardArrowDown,
-                            contentDescription = if (expanded) {
-                                stringResource(R.string.show_less)
-                            } else {
-                                stringResource(R.string.show_more)
-                            }
-                        )
-                    }
-                }
-            }
+                    .padding(start = (index * 12).dp)
+                    .size(20.dp)
+                    .clip(CircleShape),
+                contentDescription = null,
+                contentScale = ContentScale.Crop
+            )
         }
     }
 }
@@ -251,38 +196,71 @@ fun PercentageChangePill(modifier: Modifier = Modifier, change: Float, date: Str
 }
 
 @Composable
-fun AboutStockCard(modifier: Modifier = Modifier, companySummary: String, companyName: String) {
-    OutlinedCard(modifier = Modifier.padding(top = 16.dp, start = 16.dp, end = 16.dp)) {
-        Column(modifier = Modifier.padding(8.dp)) {
-            Text(
-                stringResource(R.string.about_company_name, companyName),
-                modifier = Modifier.padding(bottom = 8.dp),
-                style = MaterialTheme.typography.titleMedium
-            )
+fun AboutStockCard(modifier: Modifier, companySummary: String, companyName: String) {
+    Column(modifier = modifier) {
+        Text(
+            text = stringResource(R.string.about_company_name, companyName),
+            modifier = Modifier.padding(bottom = 8.dp),
+            style = MaterialTheme.typography.titleMedium
+        )
 
-            ExpandableText(text = companySummary, collapsedMaxLine = 4)
-        }
+        ExpandableText(
+            text = companySummary,
+            collapsedMaxLine = 4,
+            style = MaterialTheme.typography.bodyMedium,
+            showMoreText = stringResource(R.string.read_more),
+            showMoreStyle = SpanStyle(
+                textDecoration = TextDecoration.Underline,
+                fontStyle = MaterialTheme.typography.linkMedium.fontStyle,
+                fontWeight = FontWeight.W500
+            ),
+            showLessText = stringResource(R.string.read_less),
+            showLessStyle = SpanStyle(
+                textDecoration = TextDecoration.Underline,
+                fontStyle = MaterialTheme.typography.linkMedium.fontStyle,
+                fontWeight = FontWeight.W500
+            )
+        )
     }
 }
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun CompanyDetailTab(modifier: Modifier = Modifier, company: CompanyDetailRemoteResponse, onClickNews: (url: String) -> Unit) {
+fun CompanyDetailTab(modifier: Modifier, company: CompanyDetailRemoteResponse, onClickNews: (url: String) -> Unit, onClickSources: () -> Unit) {
     val titles = listOf("Overview", "Key ratios", "News")
     val selectedTabIndex = remember { mutableIntStateOf(0) }
-    Column(modifier = Modifier.fillMaxSize()) {
-        PrimaryTabRow(selectedTabIndex = selectedTabIndex.intValue) {
+    val gptInvestorColors = LocalGPTInvestorColors.current
+    Column(modifier = modifier) {
+        PrimaryTabRow(
+            modifier = Modifier
+                .padding(horizontal = 16.dp)
+                .clip(RoundedCornerShape(50)),
+            selectedTabIndex = selectedTabIndex.intValue,
+            indicator = {},
+            containerColor = MaterialTheme.colorScheme.surface
+        ) {
             titles.forEachIndexed { index, title ->
+                val isSelected = selectedTabIndex.intValue == index
+
+                val backgroundColor =
+                    if (isSelected) gptInvestorColors.utilColors.borderBright10 else Color.Transparent
+
                 Tab(
-                    selected = selectedTabIndex.intValue == index,
+                    modifier = Modifier
+                        .height(IntrinsicSize.Min)
+                        .clip(RoundedCornerShape(50))
+                        .background(backgroundColor),
+                    selected = isSelected,
+                    unselectedContentColor = gptInvestorColors.textColors.secondary50,
                     onClick = {
                         selectedTabIndex.intValue = index
                     },
                     text = {
                         Text(
                             text = title,
-                            maxLines = 2,
-                            overflow = TextOverflow.Ellipsis
+                            maxLines = 1,
+                            overflow = TextOverflow.Ellipsis,
+                            style = if (isSelected) MaterialTheme.typography.labelLarge else MaterialTheme.typography.bodyMedium
                         )
                     }
                 )
@@ -297,7 +275,12 @@ fun CompanyDetailTab(modifier: Modifier = Modifier, company: CompanyDetailRemote
             when (targetState) {
                 0 -> {
                     CompanyHistoryGraph(
-                        historicalData = company.historicalData
+                        modifier = Modifier,
+                        historicalData = company.historicalData,
+                        companySummary = company.about,
+                        companyName = company.name,
+                        sources = company.news.map { it.toPresentation() },
+                        onClickSources = onClickSources
                     )
                 }
 
@@ -323,13 +306,22 @@ fun CompanyDetailTab(modifier: Modifier = Modifier, company: CompanyDetailRemote
 }
 
 @Composable
-fun CompanyHistoryGraph(modifier: Modifier = Modifier, historicalData: List<HistoricalData>) {
+fun CompanyHistoryGraph(
+    modifier: Modifier,
+    historicalData: List<HistoricalData>,
+    companyName: String,
+    companySummary: String,
+    sources: List<NewsPresentation> = emptyList(),
+    onClickSources: () -> Unit
+) {
     Surface(
-        modifier = Modifier
+        modifier = modifier
             .fillMaxSize()
-            .padding(16.dp)
+            .padding(16.dp),
+        shape = RoundedCornerShape(16.dp),
+        border = BorderStroke(width = 2.dp, color = DividerDefaults.color)
     ) {
-        var selected by remember { mutableStateOf<TimePeriod>(TimePeriod.OneYear()) }
+        var selected by remember { mutableStateOf<TimePeriod>(TimePeriod.OneWeek()) }
 
         Column(modifier = Modifier.fillMaxWidth()) {
             Box(
@@ -372,116 +364,201 @@ fun CompanyHistoryGraph(modifier: Modifier = Modifier, historicalData: List<Hist
                 TimePeriod.ThreeMonths(),
                 TimePeriod.OneYear()
             )
-            SingleChoiceSegmentedButtonRow(
+
+            TimePeriodRow(
                 modifier = Modifier
-                    .align(Alignment.Start)
-                    .padding(start = 4.dp, end = 4.dp, bottom = 4.dp)
-            ) {
-                options.forEachIndexed { index, timePeriod ->
-                    SegmentedButton(
-                        icon = {},
-                        shape = SegmentedButtonDefaults.itemShape(
-                            index = index,
-                            count = options.size
-                        ),
-                        onClick = {
-                            selected = options[index]
-                        },
-                        selected = selected == timePeriod
-                    ) {
-                        Text(timePeriod.title)
-                    }
+                    .fillMaxWidth()
+                    .padding(top = 8.dp),
+                list = options,
+                selected = selected,
+                onClick = {
+                    selected = it
                 }
+            )
+
+            HorizontalDivider(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(horizontal = 16.dp, vertical = 16.dp),
+                thickness = 2.dp
+            )
+
+            // About company card
+            AboutStockCard(
+                modifier = Modifier.padding(start = 16.dp, end = 16.dp),
+                companySummary = companySummary,
+                companyName = companyName
+            )
+
+            CompanyDetailDataSource(
+                modifier = Modifier.padding(16.dp),
+                list = sources,
+                onClick = onClickSources
+            )
+        }
+    }
+}
+
+@Composable
+fun TimePeriodRow(modifier: Modifier, list: List<TimePeriod>, selected: TimePeriod, onClick: (period: TimePeriod) -> Unit) {
+    Row(
+        modifier = modifier,
+        verticalAlignment = Alignment.CenterVertically,
+        horizontalArrangement = Arrangement.SpaceEvenly
+    ) {
+        list.forEach {
+            SingleTimePeriod(
+                modifier = Modifier.padding(end = 8.dp),
+                period = it,
+                isSelected = it == selected,
+                onClick = onClick
+            )
+        }
+    }
+}
+
+@Composable
+fun SingleTimePeriod(modifier: Modifier, period: TimePeriod, isSelected: Boolean, onClick: (period: TimePeriod) -> Unit) {
+    if (isSelected) {
+        Surface(
+            modifier = modifier,
+            shape = RoundedCornerShape(8.dp),
+            onClick = { onClick(period) },
+            color = LocalGPTInvestorColors.current.utilColors.borderBright10
+        ) {
+            Text(
+                modifier = Modifier.padding(horizontal = 8.dp, vertical = 12.dp),
+                text = period.title,
+                style = MaterialTheme.typography.bodySmall
+            )
+        }
+    } else {
+        Text(
+            modifier = Modifier.clickable { onClick(period) },
+            text = period.title,
+            style = MaterialTheme.typography.bodySmall,
+            color = LocalGPTInvestorColors.current.textColors.secondary50
+        )
+    }
+}
+
+@Composable
+fun CompanyKeyRatios(modifier: Modifier, marketCap: Long, peRatio: Float, revenue: Long) {
+    Surface(
+        modifier = modifier
+            .fillMaxWidth()
+            .padding(16.dp),
+        shape = RoundedCornerShape(16.dp),
+        border = BorderStroke(width = 2.dp, color = DividerDefaults.color)
+    ) {
+        Column(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(16.dp),
+            verticalArrangement = Arrangement.spacedBy(8.dp)
+        ) {
+            Column(modifier = Modifier.padding()) {
+                Text(
+                    modifier = Modifier.padding(bottom = 16.dp),
+                    text = stringResource(R.string.market_cap).uppercase(),
+                    style = MaterialTheme.typography.titleMedium
+                )
+                Text(
+                    text = "$${marketCap.toReadable()}",
+                    style = MaterialTheme.typography.bodyLarge
+                )
+            }
+
+            HorizontalDivider(modifier = Modifier.fillMaxWidth(), thickness = 2.dp)
+
+            Column(modifier = Modifier.padding()) {
+                Text(
+                    modifier = Modifier.padding(bottom = 16.dp),
+                    text = stringResource(R.string.pe_ratio).uppercase(),
+                    style = MaterialTheme.typography.titleMedium
+                )
+                Text(
+                    text = "${peRatio.toTwoDecimalPlaces()}",
+                    style = MaterialTheme.typography.bodyLarge
+                )
+            }
+
+            HorizontalDivider(modifier = Modifier.fillMaxWidth(), thickness = 2.dp)
+
+            Column(modifier = Modifier.padding()) {
+                Text(
+                    modifier = Modifier.padding(bottom = 16.dp),
+                    text = stringResource(R.string.revenue).uppercase(),
+                    style = MaterialTheme.typography.titleMedium
+                )
+                Text(
+                    text = "$${revenue.toReadable()}",
+                    style = MaterialTheme.typography.bodyLarge
+                )
             }
         }
     }
 }
 
 @Composable
-fun CompanyKeyRatios(modifier: Modifier = Modifier, marketCap: Long, peRatio: Float, revenue: Long) {
-    Column(
-        modifier = Modifier
+fun CompanyDetailsNews(modifier: Modifier, news: List<NewsPresentation>, onClick: (url: String) -> Unit) {
+    Surface(
+        modifier = modifier
             .fillMaxWidth()
-            .padding(horizontal = 16.dp)
+            .padding(16.dp),
+        shape = RoundedCornerShape(16.dp)
     ) {
-        Column(modifier = Modifier.padding(top = 16.dp)) {
-            Text(
-                modifier = Modifier.padding(bottom = 16.dp),
-                text = stringResource(R.string.market_cap).uppercase(),
-                style = MaterialTheme.typography.titleMedium
-            )
-            Text("$${marketCap.toReadable()}")
-            HorizontalDivider(modifier = Modifier.fillMaxWidth())
-        }
+        Column(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(16.dp),
+            verticalArrangement = Arrangement.spacedBy(16.dp)
+        ) {
+            news.take(3).forEachIndexed { index, item ->
+                CompanyDetailNewsItem(
+                    modifier = Modifier,
+                    news = item,
+                    onClick = onClick
+                )
 
-        Column(modifier = Modifier.padding(top = 16.dp)) {
-            Text(
-                modifier = Modifier.padding(bottom = 16.dp),
-                text = stringResource(R.string.pe_ratio).uppercase(),
-                style = MaterialTheme.typography.titleMedium
-            )
-            Text("${peRatio.toTwoDecimalPlaces()}")
-            HorizontalDivider(modifier = Modifier.fillMaxWidth())
-        }
-
-        Column(modifier = Modifier.padding(top = 16.dp)) {
-            Text(
-                modifier = Modifier.padding(bottom = 16.dp),
-                text = stringResource(R.string.revenue).uppercase(),
-                style = MaterialTheme.typography.titleMedium
-            )
-            Text("$${revenue.toReadable()}")
-            HorizontalDivider(modifier = Modifier.fillMaxWidth())
-        }
-    }
-}
-
-@Composable
-fun CompanyDetailsNews(modifier: Modifier = Modifier, news: List<NewsPresentation>, onClick: (url: String) -> Unit) {
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = 16.dp)
-    ) {
-        news.take(3).forEach { item ->
-            CompanyDetailNewsItem(news = item, onClick = onClick)
-        }
-    }
-}
-
-@Composable
-fun CompanyDetailNewsItem(modifier: Modifier = Modifier, news: NewsPresentation, onClick: (url: String) -> Unit) {
-    OutlinedCard(modifier = Modifier.fillMaxWidth().padding(top = 16.dp)) {
-        Column(modifier = Modifier.padding(8.dp)) {
-            Text(
-                text = news.publisher,
-                modifier = Modifier.padding(bottom = 8.dp),
-                maxLines = 2,
-                style = MaterialTheme.typography.titleMedium,
-                overflow = TextOverflow.Ellipsis
-            )
-
-            Text(
-                modifier = Modifier.padding(bottom = 8.dp),
-                text = news.title,
-                maxLines = 2,
-                overflow = TextOverflow.Ellipsis
-            )
-
-            Column(
-                modifier = Modifier
-                    .padding(bottom = 4.dp)
-                    .clickable { onClick(Uri.encode(news.link)) }
-            ) {
-                Row {
-                    Text(stringResource(R.string.learn_more).uppercase())
-                    Image(
-                        imageVector = Icons.Default.KeyboardArrowDown,
-                        contentDescription = null
+                if (index < 2) {
+                    HorizontalDivider(
+                        modifier = Modifier
+                            .fillMaxWidth()
                     )
                 }
-                HorizontalDivider(modifier = Modifier.width(128.dp))
             }
+        }
+    }
+}
+
+@Composable
+fun CompanyDetailNewsItem(modifier: Modifier, news: NewsPresentation, onClick: (url: String) -> Unit) {
+    Column(modifier = modifier) {
+        Text(
+            text = news.publisher,
+            modifier = Modifier.padding(bottom = 8.dp),
+            maxLines = 2,
+            style = MaterialTheme.typography.titleMedium,
+            overflow = TextOverflow.Ellipsis
+        )
+
+        Text(
+            modifier = Modifier.padding(bottom = 8.dp),
+            text = news.title,
+            style = MaterialTheme.typography.bodyMedium
+        )
+
+        Column(
+            modifier = Modifier
+                .padding(bottom = 4.dp)
+                .clickable { onClick(Uri.encode(news.link)) }
+        ) {
+            Text(
+                text = stringResource(R.string.read_more),
+                style = MaterialTheme.typography.linkMedium,
+                textDecoration = TextDecoration.Underline
+            )
         }
     }
 }
@@ -492,7 +569,10 @@ fun PreviewComposable(modifier: Modifier = Modifier) {
     GPTInvestorTheme {
         Surface {
             Column(modifier = Modifier.fillMaxSize()) {
-                CompanyDetailDataSource(list = listOf(), source = "")
+                val tabs = listOf("Overview", "Key Ratio", "News", "Another Tab")
+                var selectedTabIndex by remember { mutableIntStateOf(0) }
+
+                CompanyDetailDataSource(modifier = Modifier, list = listOf(), onClick = {})
 
                 CompanyDetailPriceCard(
                     ticker = "AAPL",
@@ -502,13 +582,20 @@ fun PreviewComposable(modifier: Modifier = Modifier) {
                 )
 
                 AboutStockCard(
+                    modifier = Modifier,
                     companySummary = "I am\nthe best \n of",
                     companyName = "Microsoft corporation"
                 )
 
-                CompanyKeyRatios(marketCap = 120000000L, peRatio = 45.3F, revenue = 1010000L)
+                CompanyKeyRatios(
+                    modifier = Modifier,
+                    marketCap = 120000000L,
+                    peRatio = 45.3F,
+                    revenue = 1010000L
+                )
 
                 CompanyDetailNewsItem(
+                    modifier = Modifier,
                     news = NewsPresentation(
                         title = "The strength of the black panther has been stripped away",
                         id = "",
