@@ -1,12 +1,11 @@
 package com.thejawnpaul.gptinvestor.features.company.presentation.ui
 
-import androidx.compose.animation.animateContentSize
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.material3.OutlinedButton
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -14,8 +13,11 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.TextStyle
 import com.halilibo.richtext.commonmark.Markdown
-import com.halilibo.richtext.ui.material3.RichText
+import com.halilibo.richtext.ui.BasicRichText
+import com.halilibo.richtext.ui.RichTextThemeProvider
+import com.thejawnpaul.gptinvestor.theme.bodyChatBody
 
 const val DEFAULT_MINIMUM_TEXT_LINEE = 10
 
@@ -27,7 +29,7 @@ const val DEFAULT_MINIMUM_TEXT_LINEE = 10
  * @param text The text to be displayed.
  */
 @Composable
-fun ExpandableRichText(modifier: Modifier = Modifier, textModifier: Modifier = Modifier, text: String) {
+fun ExpandableRichText(modifier: Modifier = Modifier, textModifier: Modifier = Modifier, text: String, bodyTextStyle: TextStyle? = null) {
     // State variables to track the expanded state, clickable state.
     var isExpanded by remember { mutableStateOf(false) }
     val clickable by remember { mutableStateOf(false) }
@@ -44,27 +46,39 @@ fun ExpandableRichText(modifier: Modifier = Modifier, textModifier: Modifier = M
             }
             .then(modifier)
     ) {
-        RichText(
-            modifier = textModifier
-                .fillMaxWidth()
-                .animateContentSize()
-        ) {
-            val textToShow = if (isExpanded) {
-                text
-            } else {
-                text.substring(startIndex = 0, endIndex = text.length.div(6))
-                    .dropLastWhile { Character.isWhitespace(it) || it == '.' }
+        RichTextThemeProvider(
+            textStyleProvider = {
+                bodyTextStyle
+                    ?: MaterialTheme.typography.bodyChatBody.copy(color = MaterialTheme.colorScheme.onBackground)
             }
-
-            Markdown(content = textToShow)
-
-            OutlinedButton(onClick = {
-                isExpanded = !isExpanded
-            }, modifier = Modifier.align(Alignment.BottomEnd)) {
-                if (isExpanded) {
-                    Text(text = "Show less")
+        ) {
+            BasicRichText {
+                val textToShow = if (isExpanded) {
+                    text
                 } else {
-                    Text(text = "Show more")
+                    if (text.length < 250) {
+                        text
+                    } else {
+                        text.substring(startIndex = 0, endIndex = text.length.div(6))
+                            .dropLastWhile { Character.isWhitespace(it) || it == '.' }
+                    }
+                }
+                Markdown(content = textToShow)
+
+                if (text.length > 250) {
+                    TextButton(
+                        modifier = Modifier
+                            .align(Alignment.BottomEnd),
+                        onClick = {
+                            isExpanded = !isExpanded
+                        }
+                    ) {
+                        if (isExpanded) {
+                            Text(text = "Show less")
+                        } else {
+                            Text(text = "Show more")
+                        }
+                    }
                 }
             }
         }
