@@ -1,42 +1,72 @@
 package com.thejawnpaul.gptinvestor.features.conversation.presentation.ui
 
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.WindowInsets
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.ime
-import androidx.compose.foundation.layout.navigationBarsPadding
-import androidx.compose.foundation.layout.windowInsetsPadding
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalSoftwareKeyboardController
-import androidx.compose.ui.unit.dp
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import androidx.navigation.NavController
-import com.thejawnpaul.gptinvestor.core.navigation.Screen
 import com.thejawnpaul.gptinvestor.features.conversation.domain.model.DefaultConversation
+import com.thejawnpaul.gptinvestor.features.conversation.domain.model.DefaultPrompt
 import com.thejawnpaul.gptinvestor.features.conversation.domain.model.StructuredConversation
-import com.thejawnpaul.gptinvestor.features.conversation.presentation.viewmodel.ConversationViewModel
-import com.thejawnpaul.gptinvestor.features.investor.presentation.ui.InputBar
+import com.thejawnpaul.gptinvestor.features.conversation.presentation.state.ConversationView
+import com.thejawnpaul.gptinvestor.features.conversation.presentation.viewmodel.ConversationAction
+import com.thejawnpaul.gptinvestor.features.conversation.presentation.viewmodel.ConversationEvent
+import java.net.URLDecoder
+import java.nio.charset.StandardCharsets
 
 @Composable
-fun ConversationScreen(modifier: Modifier = Modifier, viewModel: ConversationViewModel, navController: NavController) {
-    val conversation = viewModel.conversation.collectAsStateWithLifecycle()
-    val genText = viewModel.genText.collectAsStateWithLifecycle()
-    val keyboardController = LocalSoftwareKeyboardController.current
+fun ConversationScreen(
+    modifier: Modifier,
+    state: ConversationView,
+    chatInput: String? = null,
+    title: String? = null,
+    onEvent: (ConversationEvent) -> Unit,
+    onAction: (ConversationAction) -> Unit
+) {
+    LaunchedEffect(key1 = chatInput) {
+        if (chatInput != null) {
+            val decodedChatInput = URLDecoder.decode(chatInput, StandardCharsets.UTF_8.toString())
+            if (title != null) {
+                val decodedTitle =
+                    title.let { URLDecoder.decode(it, StandardCharsets.UTF_8.toString()) }
+                onEvent(
+                    ConversationEvent.DefaultPromptClicked(
+                        prompt = DefaultPrompt(
+                            title = decodedTitle,
+                            query = decodedChatInput
+                        )
+                    )
+                )
+            } else {
+                onEvent(ConversationEvent.SendPrompt(query = decodedChatInput))
+            }
+        }
+    }
 
-    Box(modifier = Modifier.fillMaxSize()) {
-        when (conversation.value.conversation) {
+    Box(modifier = modifier.fillMaxSize()) {
+        when (state.conversation) {
             is DefaultConversation -> {
-                val default = conversation.value.conversation as DefaultConversation
+                val default = state.conversation
                 DefaultConversationScreen(
                     modifier = Modifier,
                     conversation = default,
-                    onNavigateUp = { navController.navigateUp() },
+                    onNavigateUp = {
+                        onAction(ConversationAction.OnGoBack)
+                    },
                     onPromptClicked = { prompt ->
-                        viewModel.getDefaultPromptResponse(prompt)
+                        onEvent(ConversationEvent.DefaultPromptClicked(prompt))
+                    },
+                    inputQuery = state.query,
+                    onInputQueryChanged = { input ->
+                        onEvent(ConversationEvent.UpdateInputQuery(input))
+                    },
+                    onSendClick = {
+                        onEvent(ConversationEvent.SendPrompt())
+                    },
+                    availableModels = state.availableModels,
+                    selectedModel = state.selectedModel,
+                    onModelChange = {
+                        onEvent(ConversationEvent.ModelChanged(model = it))
                     }
                 )
             }
@@ -45,40 +75,40 @@ fun ConversationScreen(modifier: Modifier = Modifier, viewModel: ConversationVie
                 // Try to pass in the VM instance down
                 StructuredConversationScreen(
                     modifier = Modifier,
-                    conversation = conversation.value.conversation as StructuredConversation,
-                    onNavigateUp = { navController.navigateUp() },
-                    text = genText.value,
+                    conversation = state.conversation,
+                    onNavigateUp = { onAction(ConversationAction.OnGoBack) },
+                    text = state.genText,
                     onClickNews = {
-                        navController.navigate(Screen.WebViewScreen.createRoute(it))
+                        onAction(ConversationAction.OnGoToWebView(url = it))
                     },
-                    onClickSuggestion = { viewModel.getSuggestedPromptResponse(it.query) }
+                    onClickFeedback = { messageId, status, reason ->
+                        onEvent(ConversationEvent.SendFeedback(messageId, status, reason))
+                    },
+                    onCopy = { text ->
+
+                        onEvent(ConversationEvent.CopyToClipboard(text))
+                    },
+                    inputQuery = state.query,
+                    onInputQueryChanged = { input ->
+                        onEvent(ConversationEvent.UpdateInputQuery(input))
+                    },
+                    onSendClick = {
+                    },
+                    companyName = "",
+                    onClickSuggestedPrompt = {
+                        onEvent(ConversationEvent.SuggestedPromptClicked(it))
+                    },
+                    availableModels = state.availableModels,
+                    selectedModel = state.selectedModel,
+                    onModelChange = {
+                        onEvent(ConversationEvent.ModelChanged(model = it))
+                    },
+                    onUpgradeModel = {}
                 )
             }
 
             else -> {
             }
         }
-
-        InputBar(
-            modifier = Modifier
-                .fillMaxWidth()
-                .align(Alignment.BottomStart)
-                .windowInsetsPadding(
-                    WindowInsets.ime
-                )
-                .navigationBarsPadding(),
-            input = conversation.value.query,
-            contentPadding = PaddingValues(0.dp),
-            sendEnabled = conversation.value.enableSend,
-            onInputChanged = { input ->
-                viewModel.updateInput(input = input)
-            },
-            onSendClick = {
-                keyboardController?.hide()
-                viewModel.getInputResponse()
-            },
-            placeholder = "Ask anything about stocks",
-            shouldRequestFocus = true
-        )
     }
 }
