@@ -2,8 +2,13 @@ package com.thejawnpaul.gptinvestor.features.conversation.presentation.ui
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.WindowInsets
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.ime
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.windowInsetsPadding
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
@@ -11,61 +16,104 @@ import androidx.compose.material.icons.automirrored.filled.ArrowBack
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
+import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalSoftwareKeyboardController
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import com.thejawnpaul.gptinvestor.R
+import com.thejawnpaul.gptinvestor.features.conversation.domain.model.AvailableModel
 import com.thejawnpaul.gptinvestor.features.conversation.domain.model.DefaultConversation
 import com.thejawnpaul.gptinvestor.features.conversation.domain.model.DefaultPrompt
+import com.thejawnpaul.gptinvestor.features.investor.presentation.ui.component.QuestionInput
 
 @Composable
-fun DefaultConversationScreen(modifier: Modifier = Modifier, conversation: DefaultConversation, onPromptClicked: (prompt: DefaultPrompt) -> Unit, onNavigateUp: () -> Unit) {
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding()
-            .verticalScroll(rememberScrollState())
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(vertical = 8.dp),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            IconButton(onClick = onNavigateUp) {
-                Icon(
-                    imageVector = Icons.AutoMirrored.Default.ArrowBack,
-                    contentDescription = stringResource(id = R.string.back)
-                )
-            }
+fun DefaultConversationScreen(
+    modifier: Modifier,
+    conversation: DefaultConversation,
+    onPromptClicked: (prompt: DefaultPrompt) -> Unit,
+    onNavigateUp: () -> Unit,
+    inputQuery: String,
+    onInputQueryChanged: (String) -> Unit,
+    onSendClick: () -> Unit,
+    availableModels: List<AvailableModel>,
+    selectedModel: AvailableModel,
+    onModelChange: (AvailableModel) -> Unit
+) {
+    val keyboardController = LocalSoftwareKeyboardController.current
 
-            Text(
-                text = stringResource(R.string.ask_gpt_investor),
-                style = MaterialTheme.typography.headlineSmall
+    Scaffold(
+        modifier = modifier.fillMaxSize(),
+        topBar = {
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(vertical = 8.dp),
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                IconButton(onClick = onNavigateUp) {
+                    Icon(
+                        imageVector = Icons.AutoMirrored.Default.ArrowBack,
+                        contentDescription = stringResource(id = R.string.back)
+                    )
+                }
+            }
+        },
+        bottomBar = {
+            QuestionInput(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .windowInsetsPadding(
+                        insets = WindowInsets.ime
+                    ),
+                onSendClicked = {
+                    keyboardController?.hide()
+                    onSendClick()
+                },
+                hint = stringResource(
+                    R.string.ask_anything_about_stocks
+                ),
+                onTextChange = { input ->
+                    onInputQueryChanged(input)
+                },
+                text = inputQuery,
+                availableModels = availableModels,
+                selectedModel = selectedModel,
+                onModelChange = {
+                    if (it.canUpgrade) {
+                        /*onEvent(
+                            HomeEvent.UpgradeModel(
+                                showBottomSheet = true,
+                                modelId = it.modelId
+                            )
+                        )
+                        return@QuestionInput*/
+                    }
+                    onModelChange(it)
+                }
             )
         }
-
-        HorizontalDivider(modifier = Modifier.fillMaxWidth())
-
-        // Ask About
-        Text(
-            text = stringResource(R.string.ask_about),
-            style = MaterialTheme.typography.bodyLarge.copy(fontWeight = FontWeight.SemiBold),
-            modifier = Modifier.padding(16.dp)
-        )
-
-        // Flow row
-        DefaultPrompts(
+    ) { innerPadding ->
+        Column(
             modifier = Modifier
-                .fillMaxWidth()
-                .padding(horizontal = 8.dp),
-            prompts = conversation.prompts,
-            onClick = onPromptClicked
-        )
+                .padding(innerPadding)
+                .fillMaxSize()
+                .verticalScroll(rememberScrollState())
+        ) {
+            HorizontalDivider(modifier = Modifier.fillMaxWidth())
+
+            Spacer(modifier = Modifier.weight(0.8f))
+
+            DefaultPrompts(
+                modifier = Modifier
+                    .weight(1.2f)
+                    .padding(horizontal = 8.dp)
+                    .align(Alignment.CenterHorizontally),
+                prompts = conversation.prompts,
+                onClick = onPromptClicked
+            )
+        }
     }
 }
