@@ -6,7 +6,7 @@ import com.thejawnpaul.gptinvestor.features.conversation.domain.model.Structured
 import kotlinx.coroutines.flow.Flow
 
 interface IHistoryRepository {
-    suspend fun getAllHistory(): Flow<Either<Failure, List<StructuredConversation>>>
+    suspend fun getAllHistory(): Flow<Either<Failure, Map<String, List<StructuredConversation>>>>
 
     suspend fun getSingleHistory(id: Long): Flow<Either<Failure, StructuredConversation>>
 }
