@@ -1,11 +1,17 @@
 package com.thejawnpaul.gptinvestor.features.investor.presentation.ui
 
+import androidx.compose.foundation.BorderStroke
+import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyRow
 import androidx.compose.foundation.lazy.items
-import androidx.compose.material3.FilterChip
+import androidx.compose.foundation.shape.CornerSize
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -13,11 +19,15 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.thejawnpaul.gptinvestor.R
 import com.thejawnpaul.gptinvestor.features.company.domain.model.SectorInput
-import com.thejawnpaul.gptinvestor.ui.theme.GPTInvestorTheme
+import com.thejawnpaul.gptinvestor.theme.GPTInvestorTheme
+import com.thejawnpaul.gptinvestor.theme.LocalGPTInvestorColors
 
 @Composable
 fun SectorChoiceQuestion(possibleAnswers: List<SectorInput>, selectedAnswer: SectorInput?, onOptionSelected: (SectorInput) -> Unit, modifier: Modifier = Modifier) {
@@ -38,26 +48,61 @@ fun SectorChoiceQuestion(possibleAnswers: List<SectorInput>, selectedAnswer: Sec
 }
 
 @Composable
-fun SingleSectorChoice(input: SectorInput, selected: Boolean, onOptionSelected: () -> Unit, modifier: Modifier = Modifier) {
-    when (input) {
+fun SingleSectorChoice(input: SectorInput, selected: Boolean, onOptionSelected: () -> Unit, modifier: Modifier) {
+    val gptInvestorColors = LocalGPTInvestorColors.current
+    val text = when (input) {
         is SectorInput.AllSector -> {
-            // Text(text = "All")
-            FilterChip(
-                modifier = Modifier,
-                selected = selected,
-                onClick = onOptionSelected,
-                label = { Text(text = "All") }
-            )
+            " All "
         }
 
-        is SectorInput.CustomSector -> {
-            // Text(text = input.sectorName)
-            FilterChip(
-                modifier = Modifier,
-                selected = selected,
-                onClick = onOptionSelected,
-                label = { Text(text = input.sectorName) }
-            )
+        else -> {
+            (input as SectorInput.CustomSector).sectorName
+        }
+    }
+
+    if (selected) {
+        Surface(
+            modifier = modifier,
+            shape = RoundedCornerShape(corner = CornerSize(20.dp)),
+            color = gptInvestorColors.utilColors.borderBright10,
+            onClick = onOptionSelected
+        ) {
+            Row(
+                modifier = Modifier.padding(horizontal = 16.dp, vertical = 12.dp),
+                horizontalArrangement = Arrangement.spacedBy(4.dp),
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                if (input.hasImage) {
+                    Image(painter = painterResource(R.drawable.ic_star), contentDescription = null)
+                }
+                Text(
+                    text = text,
+                    style = MaterialTheme.typography.labelLarge
+                )
+            }
+        }
+    } else {
+        Surface(
+            modifier = modifier,
+            shape = RoundedCornerShape(corner = CornerSize(20.dp)),
+            border = BorderStroke(width = 2.dp, color = MaterialTheme.colorScheme.outlineVariant),
+            onClick = onOptionSelected
+        ) {
+            Row(
+                modifier = Modifier.padding(horizontal = 16.dp, vertical = 12.dp),
+                horizontalArrangement = Arrangement.spacedBy(4.dp),
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                if (input.hasImage) {
+                    Image(painter = painterResource(R.drawable.ic_star), contentDescription = null)
+                }
+
+                Text(
+                    text = text,
+                    style = MaterialTheme.typography.bodyMedium,
+                    color = gptInvestorColors.textColors.secondary50
+                )
+            }
         }
     }
 }
