@@ -12,5 +12,8 @@ data class TopPickRemote(
     @field:Json(name = "key_metrics") val metrics: List<String>,
     @field:Json(name = "risks") val risks: List<String>,
     @field:Json(name = "confidence_score") val confidenceScore: Int,
-    @field:Json(name = "date") val date: String
+    @field:Json(name = "date") val date: String,
+    @field:Json(name = "price") val price: Float? = null,
+    @field:Json(name = "change") val percentageChange: Float? = null,
+    @field:Json(name = "logo_url") val imageUrl: String? = null
 )
