@@ -5,6 +5,8 @@ import com.thejawnpaul.gptinvestor.core.api.ApiService
 import com.thejawnpaul.gptinvestor.core.functional.Either
 import com.thejawnpaul.gptinvestor.core.functional.Failure
 import com.thejawnpaul.gptinvestor.core.remoteconfig.RemoteConfig
+import com.thejawnpaul.gptinvestor.core.utility.Constants
+import com.thejawnpaul.gptinvestor.features.company.data.local.dao.CompanyDao
 import com.thejawnpaul.gptinvestor.features.toppick.data.local.dao.TopPickDao
 import com.thejawnpaul.gptinvestor.features.toppick.data.local.model.TopPickEntity
 import com.thejawnpaul.gptinvestor.features.toppick.domain.model.TopPick
@@ -15,42 +17,25 @@ import java.util.Locale
 import javax.inject.Inject
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
+import kotlinx.coroutines.flow.map
 import timber.log.Timber
 
 class TopPickRepository @Inject constructor(
     private val apiService: ApiService,
     private val topPickDao: TopPickDao,
+    private val companyDao: CompanyDao,
     private val analyticsLogger: AnalyticsLogger,
     private val remoteConfig: RemoteConfig
 ) :
     ITopPickRepository {
-    override suspend fun getTopPicks(): Flow<Either<Failure, List<TopPick>>> = flow {
+    override suspend fun getTopPicks(): Flow<Either<Failure, Unit>> = flow {
         try {
             val today = SimpleDateFormat("yyyy-MM-dd", Locale.getDefault()).format(Date())
-            // emit local
-            val local = topPickDao.getAllTopPicks().filter { it.date == today }.map { entity ->
-                with(entity) {
-                    TopPick(
-                        id = id,
-                        companyName = companyName,
-                        ticker = ticker,
-                        rationale = rationale,
-                        metrics = metrics,
-                        risks = risks,
-                        confidenceScore = confidenceScore,
-                        isSaved = isSaved
-                    )
-                }
-            }.sortedByDescending { it.confidenceScore }
-
-            if (local.isNotEmpty()) {
-                Either.Right(local)
-            }
 
             val response = apiService.getTopPicks(date = today)
             if (response.isSuccessful) {
-                response.body()?.let {
-                    val newPicks = it.map { aa ->
+                response.body()?.let { remotePick ->
+                    val topPickEntities = remotePick.map { aa ->
                         with(aa) {
                             TopPickEntity(
                                 id = id,
@@ -60,33 +45,17 @@ class TopPickRepository @Inject constructor(
                                 metrics = metrics,
                                 risks = risks,
                                 confidenceScore = confidenceScore,
-                                date = date
+                                date = date,
+                                price = price ?: 0.0f,
+                                change = percentageChange ?: 0.0f,
+                                imageUrl = imageUrl ?: ""
                             )
                         }
                     }
-                    topPickDao.replaceUnsavedWithNewPicks(newPicks)
+                    topPickDao.replaceUnsavedWithNewPicks(topPickEntities)
+                    emit(Either.Right(Unit))
                 } ?: emit(Either.Left(Failure.DataError))
             }
-
-            // emit local
-            emit(
-                Either.Right(
-                    topPickDao.getAllTopPicks().filter { it.date == today }.map { entity ->
-                        with(entity) {
-                            TopPick(
-                                id = id,
-                                companyName = companyName,
-                                ticker = ticker,
-                                rationale = rationale,
-                                metrics = metrics,
-                                risks = risks,
-                                confidenceScore = confidenceScore,
-                                isSaved = isSaved
-                            )
-                        }
-                    }.sortedByDescending { it.confidenceScore }
-                )
-            )
         } catch (e: Exception) {
             Timber.e(e.stackTraceToString())
             emit(Either.Left(Failure.ServerError))
@@ -97,14 +66,17 @@ class TopPickRepository @Inject constructor(
         try {
             val pick = with(topPickDao.getSingleTopPick(pickId)) {
                 TopPick(
-                    id,
-                    companyName,
-                    ticker,
-                    rationale,
-                    metrics,
-                    risks,
-                    confidenceScore,
-                    isSaved
+                    id = id,
+                    companyName = companyName,
+                    ticker = ticker,
+                    rationale = rationale,
+                    metrics = metrics,
+                    risks = risks,
+                    confidenceScore = confidenceScore,
+                    isSaved = isSaved,
+                    imageUrl = imageUrl,
+                    percentageChange = change,
+                    currentPrice = price
                 )
             }
             emit(Either.Right(pick))
@@ -124,14 +96,17 @@ class TopPickRepository @Inject constructor(
             topPickDao.updateTopPick(entity)
             val pick = with(topPickDao.getSingleTopPick(id)) {
                 TopPick(
-                    id,
-                    companyName,
-                    ticker,
-                    rationale,
-                    metrics,
-                    risks,
-                    confidenceScore,
-                    isSaved
+                    id = id,
+                    companyName = companyName,
+                    ticker = ticker,
+                    rationale = rationale,
+                    metrics = metrics,
+                    risks = risks,
+                    confidenceScore = confidenceScore,
+                    isSaved = isSaved,
+                    imageUrl = imageUrl,
+                    percentageChange = change,
+                    currentPrice = price
                 )
             }
             emit(Either.Right(pick))
@@ -151,14 +126,17 @@ class TopPickRepository @Inject constructor(
             topPickDao.updateTopPick(entity)
             val pick = with(topPickDao.getSingleTopPick(id)) {
                 TopPick(
-                    id,
-                    companyName,
-                    ticker,
-                    rationale,
-                    metrics,
-                    risks,
-                    confidenceScore,
-                    isSaved
+                    id = id,
+                    companyName = companyName,
+                    ticker = ticker,
+                    rationale = rationale,
+                    metrics = metrics,
+                    risks = risks,
+                    confidenceScore = confidenceScore,
+                    isSaved = isSaved,
+                    imageUrl = imageUrl,
+                    percentageChange = change,
+                    currentPrice = price
                 )
             }
             emit(Either.Right(pick))
@@ -171,7 +149,7 @@ class TopPickRepository @Inject constructor(
     override suspend fun shareTopPick(id: String): Flow<Either<Failure, String>> = flow {
         try {
             val pick = topPickDao.getSingleTopPick(id)
-            val domain = remoteConfig.fetchAndActivateStringValue("website_domain")
+            val domain = remoteConfig.fetchAndActivateStringValue(Constants.WEBSITE_DOMAIN_KEY)
 
             val urlToShare = "${domain}single-pick/${pick.id}"
 
@@ -213,7 +191,11 @@ class TopPickRepository @Inject constructor(
                                 metrics = metrics,
                                 risks = risks,
                                 confidenceScore = confidenceScore,
-                                isSaved = isSaved
+                                isSaved = isSaved,
+                                imageUrl = imageUrl,
+                                percentageChange = change,
+                                currentPrice = price
+
                             )
                         }
                     }
@@ -237,10 +219,68 @@ class TopPickRepository @Inject constructor(
                     metrics = metrics,
                     risks = risks,
                     confidenceScore = confidenceScore,
-                    isSaved = isSaved
+                    isSaved = isSaved,
+                    imageUrl = imageUrl,
+                    percentageChange = change,
+                    currentPrice = price
                 )
             }
         }.sortedByDescending { it.confidenceScore }
         emit(Either.Right(local))
+    }
+
+    override suspend fun getTopPicksByDate(): Flow<List<TopPick>> {
+        val today = SimpleDateFormat("yyyy-MM-dd", Locale.getDefault()).format(Date())
+        return try {
+            topPickDao.getTopPicksFlow(today).map { list ->
+                list.map { entity ->
+                    with(entity) {
+                        TopPick(
+                            id = id,
+                            companyName = companyName,
+                            ticker = ticker,
+                            rationale = rationale,
+                            metrics = metrics,
+                            risks = risks,
+                            confidenceScore = confidenceScore,
+                            isSaved = isSaved,
+                            imageUrl = imageUrl,
+                            percentageChange = change,
+                            currentPrice = price
+                        )
+                    }
+                }
+            }
+        } catch (e: Exception) {
+            Timber.e(e.stackTraceToString())
+            flow { emit(emptyList()) }
+        }
+    }
+
+    override suspend fun searchTopPicks(query: String): Flow<List<TopPick>> {
+        return try {
+            topPickDao.searchTopPicks(query).map { list ->
+                list.map { entity ->
+                    with(entity) {
+                        TopPick(
+                            id = id,
+                            companyName = companyName,
+                            ticker = ticker,
+                            rationale = rationale,
+                            metrics = metrics,
+                            risks = risks,
+                            confidenceScore = confidenceScore,
+                            isSaved = isSaved,
+                            imageUrl = imageUrl,
+                            percentageChange = change,
+                            currentPrice = price
+                        )
+                    }
+                }
+            }
+        } catch (e: Exception) {
+            Timber.e(e.stackTraceToString())
+            flow { emit(emptyList()) }
+        }
     }
 }
