@@ -4,7 +4,6 @@ import com.thejawnpaul.gptinvestor.core.baseusecase.BaseUseCase
 import com.thejawnpaul.gptinvestor.core.di.IoDispatcher
 import com.thejawnpaul.gptinvestor.core.functional.Either
 import com.thejawnpaul.gptinvestor.core.functional.Failure
-import com.thejawnpaul.gptinvestor.features.toppick.domain.model.TopPick
 import com.thejawnpaul.gptinvestor.features.toppick.domain.repository.ITopPickRepository
 import javax.inject.Inject
 import kotlinx.coroutines.CoroutineDispatcher
@@ -15,11 +14,11 @@ class GetTopPicksUseCase @Inject constructor(
     @IoDispatcher private val dispatcher: CoroutineDispatcher,
     coroutineScope: CoroutineScope,
     private val repository: ITopPickRepository
-) : BaseUseCase<GetTopPicksUseCase.None, List<TopPick>>(coroutineScope, dispatcher) {
+) : BaseUseCase<GetTopPicksUseCase.None, Unit>(coroutineScope, dispatcher) {
 
     class None
 
-    override suspend fun run(params: None): Flow<Either<Failure, List<TopPick>>> {
+    override suspend fun run(params: None): Flow<Either<Failure, Unit>> {
         return repository.getTopPicks()
     }
 }
