@@ -4,14 +4,14 @@ import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.material3.Button
+import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.LinearProgressIndicator
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedCard
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -19,12 +19,14 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextDecoration
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.thejawnpaul.gptinvestor.R
 import com.thejawnpaul.gptinvestor.features.toppick.presentation.model.TopPickPresentation
 import com.thejawnpaul.gptinvestor.features.toppick.presentation.state.TopPicksView
-import com.thejawnpaul.gptinvestor.ui.theme.GPTInvestorTheme
+import com.thejawnpaul.gptinvestor.theme.GPTInvestorTheme
+import com.thejawnpaul.gptinvestor.theme.linkMedium
 
 @Composable
 fun TopPicks(modifier: Modifier = Modifier, state: TopPicksView, onClickRetry: () -> Unit, onClick: (id: String) -> Unit, onClickSeeAll: () -> Unit) {
@@ -33,7 +35,7 @@ fun TopPicks(modifier: Modifier = Modifier, state: TopPicksView, onClickRetry: (
             // Text
             Text(
                 text = stringResource(R.string.top_picks),
-                style = MaterialTheme.typography.headlineMedium.copy(fontWeight = FontWeight.SemiBold),
+                style = MaterialTheme.typography.titleMedium.copy(fontWeight = FontWeight.SemiBold),
                 modifier = Modifier.padding(16.dp)
             )
 
@@ -61,23 +63,35 @@ fun TopPicks(modifier: Modifier = Modifier, state: TopPicksView, onClickRetry: (
                 // Text
                 Text(
                     text = stringResource(R.string.top_picks),
-                    style = MaterialTheme.typography.headlineMedium.copy(fontWeight = FontWeight.SemiBold),
+                    style = MaterialTheme.typography.titleMedium.copy(fontWeight = FontWeight.SemiBold),
                     modifier = Modifier.padding(16.dp)
                 )
 
                 // Text
                 Text(
                     text = stringResource(R.string.see_all),
-                    style = MaterialTheme.typography.titleMedium,
+                    style = MaterialTheme.typography.linkMedium,
                     modifier = Modifier
                         .padding(end = 16.dp)
-                        .clickable(onClick = onClickSeeAll)
+                        .clickable(onClick = onClickSeeAll),
+                    textDecoration = TextDecoration.Underline
                 )
             }
 
-            state.topPicks.take(2).forEach {
-                SingleTopPickItem(pickPresentation = it, onClick = onClick)
-                Spacer(modifier = Modifier.size(16.dp))
+            OutlinedCard(modifier = Modifier.padding(horizontal = 16.dp)) {
+                val numberOfItemsToShow = 2
+                Column(modifier = Modifier.padding(horizontal = 16.dp)) {
+                    state.topPicks.take(numberOfItemsToShow).forEachIndexed { index, presentation ->
+                        HomeTopPickItem(
+                            modifier = Modifier.padding(vertical = 16.dp),
+                            pickPresentation = presentation,
+                            onClick = onClick
+                        )
+                        if (index < numberOfItemsToShow - 1) {
+                            HorizontalDivider(modifier = Modifier.fillMaxWidth())
+                        }
+                    }
+                }
             }
         }
     }
@@ -102,7 +116,10 @@ fun TopPicksPreview(modifier: Modifier = Modifier) {
                                 confidenceScore = 2,
                                 metrics = emptyList(),
                                 risks = emptyList(),
-                                isSaved = true
+                                isSaved = true,
+                                imageUrl = "",
+                                percentageChange = 0.0f,
+                                currentPrice = 0.0f
                             ),
                             TopPickPresentation(
                                 id = "2",
@@ -112,7 +129,10 @@ fun TopPicksPreview(modifier: Modifier = Modifier) {
                                 confidenceScore = 2,
                                 metrics = emptyList(),
                                 risks = emptyList(),
-                                isSaved = false
+                                isSaved = false,
+                                imageUrl = "",
+                                percentageChange = 0.0f,
+                                currentPrice = 0.0f
                             ),
                             TopPickPresentation(
                                 id = "2",
@@ -122,7 +142,10 @@ fun TopPicksPreview(modifier: Modifier = Modifier) {
                                 confidenceScore = 2,
                                 metrics = emptyList(),
                                 risks = emptyList(),
-                                isSaved = true
+                                isSaved = true,
+                                imageUrl = "",
+                                percentageChange = 0.0f,
+                                currentPrice = 0.0f
                             )
                         )
                     ),
